# openapi/main.yaml
openapi: 3.0.3
info:
  title: "Comment Rest API"
  description: "All comment Rest API Endpoints"
  version: "v1"
servers:
  - url: "https"
  - url: 'http'
tags:
  - name: "Demo"
    description: "Everything about Comment"
paths:
  /comment:
    post:
      summary: "add new comment"
      operationId: "addComment"
      tags:
        - "Comment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentDTO'
      responses:
        201:
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentIdDTO'
    get:
      summary: "Get all comments"
      operationId: "getAllComments"
      tags:
        - "Comment"
      responses:
        '200':
          description: "Successful response"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CommentDTO"
  /comment/{commentId}/post:
    put:
      summary: "post comment"
      description: "post comment"
      operationId: "postComment"
      tags:
        - "Comment"
      parameters:
        - name: "commentId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        '204':
          description: "Successful operation"
  /comment/{commentId}/remove:
    put:
      summary: "remove comment"
      description: "remove comment"
      operationId: "removeComment"
      tags:
        - "Comment"
      parameters:
        - name: "commentId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        '204':
          description: "Successful operation"
  /comment/{commentId}:
    get:
      summary: "get comment by Id"
      operationId: "getCommentById"
      tags:
        - "Comment"
      parameters:
        - name: "commentId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        '200':
          description: "Successful response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentDTO"
components:
  schemas:
    CommentDTO:
      properties:
        title:
          type: "string"
        content:
          type: "string"
    AddCommentDTO:
      properties:
        title:
          type: "string"
        content:
          type: "string"
    CommentIdDTO:
      properties:
        value:
          type: string
          format: uuid